requirements.txt
Django==2.2
django-cleanup==3.2.0
django-crispy-forms==1.7.2
mysql-connector-python==8.0.15
Pillow==6.0.0
pytz==2018.9
sqlparse==0.3.0
virtualenv==16.4.3
virtualenvwrapper-win==1.2.5
